update:package-name
ディレクトリ（パッケージ）を
main/java/com/example/app
から
サービス名を用いたものに変更する
