--- app/src/main/java/com/example/app/AppApplication.java
@@ -1,12 +1,0 @@
-package com.example.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AppApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(AppApplication.class, args);
-  }
-}
