--- app/src/main/java/com/example/app/application/package-info.java
@@ -1,3 +1,0 @@
-package com.example.app.application;
-
-/** Application services, use cases, and transactional orchestration. */
