--- app/src/main/java/com/example/app/domain/package-info.java
@@ -1,3 +1,0 @@
-package com.example.app.domain;
-
-/** Domain entities, value objects, and domain services. */
