--- app/src/main/java/com/example/app/domain/user/User.java
@@ -1,6 +1,0 @@
-package com.example.app.domain.user;
-
-import java.time.Instant;
-
-/** ユーザーのドメインモデル。Auth0 の subject と作成・更新時刻のみを保持する。 */
-public record User(String auth0Id, Instant createdAt, Instant updatedAt) {}
