--- app/src/main/java/com/example/app/infrastructure/external/package-info.java
@@ -1,3 +1,0 @@
-package com.example.app.infrastructure.external;
-
-/** Clients for Auth0 and other external service integrations. */
