--- app/src/main/java/com/example/app/infrastructure/persistence/package-info.java
@@ -1,3 +1,0 @@
-package com.example.app.infrastructure.persistence;
-
-/** Persistence adapters, JPA entities, and repository implementations. */
