--- app/src/main/java/com/example/app/infrastructure/storage/package-info.java
@@ -1,3 +1,0 @@
-package com.example.app.infrastructure.storage;
-
-/** Object storage integrations for Cloudflare R2 compatible APIs. */
