--- app/src/main/java/com/example/app/interfaces/api/package-info.java
@@ -1,3 +1,0 @@
-package com.example.app.interfaces.api;
-
-/** REST controllers, request/response DTOs, and exception handlers. */
