--- app/src/main/java/com/example/app/interfaces/config/package-info.java
@@ -1,3 +1,0 @@
-package com.example.app.interfaces.config;
-
-/** Application-facing configuration such as security, CORS, and OpenAPI. */
