+++ app/src/main/java/io/github/tempsotsusei/kotobanotane/AppApplication.java
@@ -1,0 +1,12 @@
+package io.github.tempsotsusei.kotobanotane;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AppApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(AppApplication.class, args);
+  }
+}
