+++ app/src/main/java/io/github/tempsotsusei/kotobanotane/application/package-info.java
@@ -1,0 +1,3 @@
+package io.github.tempsotsusei.kotobanotane.application;
+
+/** Application services, use cases, and transactional orchestration. */
