+++ app/src/main/java/io/github/tempsotsusei/kotobanotane/domain/package-info.java
@@ -1,0 +1,3 @@
+package io.github.tempsotsusei.kotobanotane.domain;
+
+/** Domain entities, value objects, and domain services. */
