+++ app/src/main/java/io/github/tempsotsusei/kotobanotane/infrastructure/external/package-info.java
@@ -1,0 +1,3 @@
+package io.github.tempsotsusei.kotobanotane.infrastructure.external;
+
+/** Clients for Auth0 and other external service integrations. */
