+++ app/src/main/java/io/github/tempsotsusei/kotobanotane/infrastructure/persistence/package-info.java
@@ -1,0 +1,3 @@
+package io.github.tempsotsusei.kotobanotane.infrastructure.persistence;
+
+/** Persistence adapters, JPA entities, and repository implementations. */
