+++ app/src/main/java/io/github/tempsotsusei/kotobanotane/infrastructure/storage/package-info.java
@@ -1,0 +1,3 @@
+package io.github.tempsotsusei.kotobanotane.infrastructure.storage;
+
+/** Object storage integrations for Cloudflare R2 compatible APIs. */
