+++ app/src/main/java/io/github/tempsotsusei/kotobanotane/interfaces/api/package-info.java
@@ -1,0 +1,3 @@
+package io.github.tempsotsusei.kotobanotane.interfaces.api;
+
+/** REST controllers, request/response DTOs, and exception handlers. */
