+++ app/src/main/java/io/github/tempsotsusei/kotobanotane/interfaces/config/package-info.java
@@ -1,0 +1,3 @@
+package io.github.tempsotsusei.kotobanotane.interfaces.config;
+
+/** Application-facing configuration such as security, CORS, and OpenAPI. */
