--- app/src/test/java/com/example/app/AppApplicationTests.java
@@ -1,11 +1,0 @@
-package com.example.app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AppApplicationTests {
-
-  @Test
-  void contextLoads() {}
-}
