+++ app/src/test/java/io/github/tempsotsusei/kotobanotane/AppApplicationTests.java
@@ -1,0 +1,11 @@
+package io.github.tempsotsusei.kotobanotane;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AppApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
